SDK-691: Fixed init, re-init setting, fixed test
