+++ src/Extension/Settings/settings.yaml
@@ -1,4 +1,18 @@
 settings:
+  - path: developer_email
+    initialization_description: What is your email?
+    strategy: overwrite
+    init: true
+    is_project: false
+    type: string
+    values: ~
+  - path: developer_github_account
+    initialization_description: What is your github account if you have the one?
+    strategy: overwrite
+    init: true
+    is_project: false
+    type: string
+    values: ~
   - path: extension_dirs
     initialization_description: Where can extensions be loaded from?
     strategy: merge
@@ -112,13 +126,4 @@ settings:
     type: array
     initializer: project_workflow_initializer
     values: []
-  - path: repository
-    initialization_description: What is private repository url?
-    strategy: overwrite
-    init: false
-    is_project: false
-    type: array
-    values:
-      - git@github.com:spryker-sdk/sdk-tasks-bundle.git
-      - feature/sdk-581-init-bundle
